finance-agents/data-agent/pyproject.toml
[project]
name = "data-agent"
version = "0.1.0"
description = "Financial data reconciliation and preparation agent powered by LangGraph"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.4.1",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "psycopg2-binary>=2.9.0",
    "chardet>=5.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
